+++ src/marshmallow/schema.py
@@ -380,6 +380,15 @@ class BaseSchema(base.SchemaABC):
         self._normalize_nested_options()
         #: Dictionary mapping field_names -> :class:`Field` objects
         self.fields = self._init_fields()
+        self.dump_fields, self.load_fields = self.dict_class(), self.dict_class()
+        for field_name, field_obj in self.fields.items():
+            if field_obj.load_only:
+                self.load_fields[field_name] = field_obj
+            elif field_obj.dump_only:
+                self.dump_fields[field_name] = field_obj
+            else:
+                self.load_fields[field_name] = field_obj
+                self.dump_fields[field_name] = field_obj
         messages = {}
         messages.update(self._default_error_messages)
         for cls in reversed(self.__class__.__mro__):
@@ -449,70 +458,24 @@ class BaseSchema(base.SchemaABC):
             return error.valid_data or missing
         return value
 
-    def _serialize(
-        self,
-        obj,
-        fields_dict,
-        *,
-        error_store,
-        many=False,
-        accessor=None,
-        dict_class=dict,
-        index_errors=True,
-        index=None
-    ):
-        """Takes raw data (a dict, list, or other object) and a dict of
-        fields to output and serializes the data based on those fields.
+    def _serialize(self, obj, *, many=False):
+        """Serialize ``obj``.
 
-        :param obj: The actual object(s) from which the fields are taken from
-        :param dict fields_dict: Mapping of field names to :class:`Field` objects.
-        :param ErrorStore error_store: Structure to store errors.
-        :param bool many: Set to `True` if ``data`` should be serialized as
-            a collection.
-        :param callable accessor: Function to use for getting values from ``obj``.
-        :param type dict_class: Dictionary class used to construct the output.
-        :param bool index_errors: Whether to store the index of invalid items in
-            ``self.errors`` when ``many=True``.
-        :param int index: Index of the item being serialized (for storing errors) if
-            serializing a collection, otherwise `None`.
-        :return: A dictionary of the marshalled data
+        :param obj: The object(s) to serialize.
+        :param bool many: `True` if ``data`` should be serialized as a collection.
+        :return: A dictionary of the serialized data
 
         .. versionchanged:: 1.0.0
             Renamed from ``marshal``.
         """
-        index = index if index_errors else None
         if many and obj is not None:
-            self._pending = True
-            ret = [
-                self._serialize(
-                    d,
-                    fields_dict,
-                    error_store=error_store,
-                    many=False,
-                    dict_class=dict_class,
-                    accessor=accessor,
-                    index=idx,
-                    index_errors=index_errors,
-                )
-                for idx, d in enumerate(obj)
-            ]
-            self._pending = False
-            return ret
+            return [self._serialize(d, many=False) for d in obj]
         ret = self.dict_class()
-        for attr_name, field_obj in fields_dict.items():
-            if getattr(field_obj, "load_only", False):
-                continue
-            key = field_obj.data_key or attr_name
-            getter = lambda d: field_obj.serialize(attr_name, d, accessor=accessor)
-            value = self._call_and_store(
-                getter_func=getter,
-                data=obj,
-                field_name=key,
-                error_store=error_store,
-                index=index,
-            )
+        for attr_name, field_obj in self.dump_fields.items():
+            value = field_obj.serialize(attr_name, obj, accessor=self.get_attribute)
             if value is missing:
                 continue
+            key = field_obj.data_key or attr_name
             ret[key] = value
         return ret
 
@@ -531,48 +494,26 @@ class BaseSchema(base.SchemaABC):
             This method returns the serialized data rather than a ``(data, errors)`` duple.
             A :exc:`ValidationError <marshmallow.exceptions.ValidationError>` is raised
             if ``obj`` is invalid.
+        .. versionchanged:: 3.0.0rc9
+            Validation no longer occurs upon serialization.
         """
-        error_store = ErrorStore()
-        errors = {}
         many = self.many if many is None else bool(many)
         if many and is_iterable_but_not_string(obj):
             obj = list(obj)
 
         if self._has_processors(PRE_DUMP):
-            try:
-                processed_obj = self._invoke_dump_processors(
-                    PRE_DUMP, obj, many=many, original_data=obj
-                )
-            except ValidationError as error:
-                errors = error.normalized_messages()
-                result = None
+            processed_obj = self._invoke_dump_processors(
+                PRE_DUMP, obj, many=many, original_data=obj
+            )
         else:
             processed_obj = obj
 
-        if not errors:
-            result = self._serialize(
-                processed_obj,
-                fields_dict=self.fields,
-                error_store=error_store,
-                many=many,
-                accessor=self.get_attribute,
-                dict_class=self.dict_class,
-                index_errors=self.opts.index_errors,
-            )
-            errors = error_store.errors
+        result = self._serialize(processed_obj, many=many)
 
-        if not errors and self._has_processors(POST_DUMP):
-            try:
-                result = self._invoke_dump_processors(
-                    POST_DUMP, result, many=many, original_data=obj
-                )
-            except ValidationError as error:
-                errors = error.normalized_messages()
-        if errors:
-            exc = ValidationError(errors, data=obj, valid_data=result)
-            # User-defined error handler
-            self.handle_error(exc, obj, many=many)
-            raise exc
+        if self._has_processors(POST_DUMP):
+            result = self._invoke_dump_processors(
+                POST_DUMP, result, many=many, original_data=obj
+            )
 
         return result
 
@@ -595,73 +536,50 @@ class BaseSchema(base.SchemaABC):
         return self.opts.render_module.dumps(serialized, *args, **kwargs)
 
     def _deserialize(
-        self,
-        data,
-        fields_dict,
-        *,
-        error_store,
-        many=False,
-        partial=False,
-        unknown=RAISE,
-        dict_class=dict,
-        index_errors=True,
-        index=None
+        self, data, *, error_store, many=False, partial=False, unknown=RAISE, index=None
     ):
-        """Deserialize ``data`` based on the schema defined by ``fields_dict``.
+        """Deserialize ``data``.
 
         :param dict data: The data to deserialize.
-        :param dict fields_dict: Mapping of field names to :class:`Field` objects.
         :param ErrorStore error_store: Structure to store errors.
-        :param bool many: Set to `True` if ``data`` should be deserialized as
-            a collection.
+        :param bool many: `True` if ``data`` should be deserialized as a collection.
         :param bool|tuple partial: Whether to ignore missing fields and not require
             any fields declared. Propagates down to ``Nested`` fields as well. If
             its value is an iterable, only missing fields listed in that iterable
             will be ignored. Use dot delimiters to specify nested fields.
         :param unknown: Whether to exclude, include, or raise an error for unknown
             fields in the data. Use `EXCLUDE`, `INCLUDE` or `RAISE`.
-        :param type dict_class: Dictionary class used to construct the output.
-        :param bool index_errors: Whether to store the index of invalid items in
-            ``self.errors`` when ``many=True``.
         :param int index: Index of the item being serialized (for storing errors) if
             serializing a collection, otherwise `None`.
         :return: A dictionary of the deserialized data.
         """
+        index_errors = self.opts.index_errors
         index = index if index_errors else None
         if many:
             if not is_collection(data):
                 error_store.store_error([self.error_messages["type"]], index=index)
                 ret = []
             else:
-                self._pending = True
                 ret = [
                     self._deserialize(
                         d,
-                        fields_dict,
                         error_store=error_store,
                         many=False,
                         partial=partial,
                         unknown=unknown,
-                        dict_class=dict_class,
                         index=idx,
-                        index_errors=index_errors,
                     )
                     for idx, d in enumerate(data)
                 ]
-                self._pending = False
             return ret
-        ret = dict_class()
+        ret = self.dict_class()
         # Check data is a dict
         if not isinstance(data, Mapping):
             error_store.store_error([self.error_messages["type"]], index=index)
         else:
             partial_is_collection = is_collection(partial)
-            for attr_name, field_obj in fields_dict.items():
-                if field_obj.dump_only:
-                    continue
-                field_name = attr_name
-                if field_obj.data_key:
-                    field_name = field_obj.data_key
+            for attr_name, field_obj in self.load_fields.items():
+                field_name = field_obj.data_key or attr_name
                 raw_value = data.get(field_name, missing)
                 if raw_value is missing:
                     # Ignore missing field if we're allowed to.
@@ -691,13 +609,12 @@ class BaseSchema(base.SchemaABC):
                     index=index,
                 )
                 if value is not missing:
-                    key = fields_dict[attr_name].attribute or attr_name
+                    key = field_obj.attribute or attr_name
                     set_value(ret, key, value)
             if unknown != EXCLUDE:
                 fields = {
                     field_obj.data_key or field_name
-                    for field_name, field_obj in fields_dict.items()
-                    if not field_obj.dump_only
+                    for field_name, field_obj in self.load_fields.items()
                 }
                 for key in set(data) - fields:
                     value = data[key]
@@ -768,7 +685,6 @@ class BaseSchema(base.SchemaABC):
         output,
         *,
         original_data,
-        fields_dict,
         error_store,
         many,
         partial,
@@ -847,13 +763,10 @@ class BaseSchema(base.SchemaABC):
             # Deserialize data
             result = self._deserialize(
                 processed_data,
-                fields_dict=self.fields,
                 error_store=error_store,
                 many=many,
                 partial=partial,
                 unknown=unknown,
-                dict_class=self.dict_class,
-                index_errors=self.opts.index_errors,
             )
             # Run field-level validation
             self._invoke_field_validators(
@@ -1025,7 +938,7 @@ class BaseSchema(base.SchemaABC):
                 field_obj.dump_only = True
             field_obj._bind_to_schema(field_name, self)
             self.on_bind_field(field_name, field_obj)
-        except TypeError as exc:
+        except TypeError as error:
             # field declared as a class, not an instance
             if isinstance(field_obj, type) and issubclass(field_obj, base.FieldABC):
                 msg = (
@@ -1033,7 +946,7 @@ class BaseSchema(base.SchemaABC):
                     "Field instance, not a class. "
                     'Did you mean "fields.{}()"?'.format(field_name, field_obj.__name__)
                 )
-                raise TypeError(msg) from exc
+                raise TypeError(msg) from error
 
     @lru_cache(maxsize=8)
     def _has_processors(self, tag):
@@ -1080,12 +993,12 @@ class BaseSchema(base.SchemaABC):
 
             try:
                 field_obj = self.fields[field_name]
-            except KeyError as exc:
+            except KeyError as error:
                 if field_name in self.declared_fields:
                     continue
                 raise ValueError(
                     '"{}" field does not exist.'.format(field_name)
-                ) from exc
+                ) from error
 
             if many:
                 for idx, item in enumerate(data):
@@ -1144,7 +1057,6 @@ class BaseSchema(base.SchemaABC):
                         validator,
                         item,
                         original_data=orig,
-                        fields_dict=self.fields,
                         error_store=error_store,
                         many=many,
                         partial=partial,
@@ -1156,7 +1068,6 @@ class BaseSchema(base.SchemaABC):
                     validator,
                     data,
                     original_data=original_data,
-                    fields_dict=self.fields,
                     error_store=error_store,
                     many=many,
                     pass_original=pass_original,